Split Marketing plans display between files
